requirements.txt
textual
playwright
google-generativeai
PyGithub
httpx[socks]
pydantic
PySocks
python-dotenv
